Changed BlobReference to use pointers instead of indices into the Blob heap.
Not all BlobReference objects point to the blob heap.  For example, some some
signatures are instantiated into temporary (or semi-permanent) buffers, and are
treated like blobs.  Therefore, since a blob can be anywhere in memory, we need
to use pointers to reference them.

This is still a bit messy, but at least the AssemblyDumpCPP test stub runs to
completion without assert, execption, or crash :-)