Cross-assembly instantiation fixes
The ElementContexts code did not correctly handle cross-module instantiation.  Normal signatures do
not contain assembly information, since they refer only to elements in the same assembly.  However,
an instantiated signature may refer to elements in other assemblies, notably the assembly in which
the instantiating element is defined.  We now have a special internal signature item that encodes
the pointer to the database in which the type entity is defined.

ElementContexts is now a mess, and we have a lot of cleanup to do, but at least it's a step forward.