Major rework of projects and solution to fix ODR violations caused by mismatch of /ZW
It turns out that <future> depends on ConcRT headers that are /ZW-sensitive.  So, we've added new
build configurations so that we can build CxxReflect both with and without /ZW support.  We've also
added some #pragma detect_mismatch directives to ensure that we don't mismatch object files.

We've also added ILoaderConfiguration to allow us to perform namespace transformations, required for
WinRT (e.g., to transform System -> Platform).

We've started reorganizing the WindowsRuntime.cpp file to move basic functionality into the loader
context so that there's less reliance on global state.  Now that we have to compile with /ZW enabled
we may want to consider removing most of the rest of the global state and exposing the class itself
via the header.