Mr. Profiler informed me that the binary searches required to find the owners of various types of rows were extremely expensive.  Ridiculously expensive, even.
The root of the problem was that we were instantiating a Row object for each step in the binary search, so for, say, 35,000 rows, we'd instantiate over 500,000 objects.  We were also calling back into the root database many times for every call, which required a lot of indirection and overhead that was not eliminatable.

This change adds a new iterator to the Database layer, a StrideIterator, that allows us to iterate over a table as an array of bytes.  This provides substantially better performance (looking up the parent MethodDef rows of each of the Param rows in mscorlib took 1.3 seconds before; it takes 10ms now).