This change fixes several bugs in signature instantitaion and member computation.
During signature instantiation, we were scope-tagging Class types but not GenericInst types.  This causes invalid signature instantiation when we are instantitaing a GenericInst and we substitute an argument with another GenericInst that comes from a different scope.

The second issue was in how we handled instantiated signatures in the membership logic.  We made the assumption that the scope of any instantiated signature was the same as the instantiating type.  This is incorrect:  the scope may be any of a number of different scopes, depending on where the original, uninstantiated signature came from.  This change fixes the membership logic to handle instantiated signatures as blobs instead of byte arrays, since blobs track their scope.