This change cleans up resolve_member_ref so that it no longer computes member tables through the membership collections.
The signature of a member_ref will always _exactly_ match the signature of the definition of the member, so for generic members and members of generics we need to use the uninstantiated signature in comparisons.  This means that we can simply resolve the parent type and iterate over its declared members.  This saves us a lot of hassle and should prove much faster.